requirements.txt
Django==2.1
psycopg2
pyaml
djangorestframework
django-filter
